+++ examples/distributed-mandelbrot/server.py
@@ -1,16 +1,6 @@
 from __future__ import print_function, division
-import math
+import sys
 import Pyro4
-
-
-# A note about the abs(z) calls below not using abs(z),
-# but instead squaring the imaginary and real components itself:
-#
-# This is because using abs(z) triggers a performance issue on pypy on windows,
-# where it is much slower than it could have been. This seems to be an issue
-# with the hypot() function in Microsoft's 32 bits runtime library.
-# See bug report https://bitbucket.org/pypy/pypy/issues/2401
-# The problem doesn't occur on other Pypy implementations.
 
 
 @Pyro4.expose
@@ -40,7 +30,7 @@ class Mandelbrot(object):
     def iterations(self, z):
         c = z
         for n in range(self.maxiters):
-            if z.real*z.real + z.imag*z.imag > 4:      # abs(z) > 2
+            if abs(z) > 2:
                 return n
             z = z*z + c
         return self.maxiters
@@ -63,26 +53,31 @@ class MandelbrotColorPixels(object):
         zi *= res_y/res_x  # aspect correction
         z = complex(zr, zi)
         c = z
+        iters = 0
         for iters in range(self.maxiters+1):
-            if z.real*z.real + z.imag*z.imag > 4:      # abs(z) > 2
+            if abs(z) > 2:
                 break
             z = z*z + c
         if iters >= self.maxiters:
             return 0, 0, 0
-        abs_z = math.sqrt(z.real*z.real + z.imag*z.imag)     # abs(z)
         r = (iters+32) % 255
-        g = (iters - math.log(abs_z)) % 255
-        b = (abs_z*iters) % 255
+        g = iters % 255
+        b = (iters+40) % 255
         return int(r), int(g), int(b)
 
 
 if __name__ == "__main__":
+    # spawn a Pyro daemon process
+    # (can't use threads, because of the GIL)
+    if len(sys.argv) != 2:
+        raise SystemExit("give argument: server_id number")
+
+    server_id = int(sys.argv[1])
     with Pyro4.Daemon() as d:
         with Pyro4.locateNS() as ns:
-            for num in range(8):
-                mandel_server = d.register(Mandelbrot())
-                mandel_color_server = d.register(MandelbrotColorPixels())
-                ns.register("mandelbrot_"+str(num), mandel_server, safe=True, metadata={"class:mandelbrot_calc"})
-                ns.register("mandelbrot_color"+str(num), mandel_color_server, safe=True, metadata={"class:mandelbrot_calc_color"})
-        print("Mandelbrot calculation server ready.")
+            mandel_server = d.register(Mandelbrot)
+            mandel_color_server = d.register(MandelbrotColorPixels)
+            ns.register("mandelbrot_"+str(server_id), mandel_server, safe=True, metadata={"class:mandelbrot_calc"})
+            ns.register("mandelbrot_color_"+str(server_id), mandel_color_server, safe=True, metadata={"class:mandelbrot_calc_color"})
+        print("Mandelbrot calculation server #{} ready.".format(server_id))
         d.requestLoop()